[MIL-1] refactor: move backend-related files
