+++ package.json
@@ -4,7 +4,7 @@
   "description": "show programming language based on the repos of github's account",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon src/app.js | gulp sass",
+    "start": "nodemon src/index.js | gulp sass",
     "lint": "eslint '**/*.js' --ignore-pattern node_modules/",
     "mocha": "mocha",
     "test": "nyc mocha"
